Added some final changes - Illegal argument exception

+++ ddtroops-persistence/src/main/java/cz/fi/muni/pa165/ddtroops/dao/impl/TroopDaoImpl.java
@@ -1,5 +1,6 @@
-package cz.fi.muni.pa165.ddtroops.dao;
+package cz.fi.muni.pa165.ddtroops.dao.impl;
 
+import cz.fi.muni.pa165.ddtroops.dao.TroopDao;
 import cz.fi.muni.pa165.ddtroops.entity.Troop;
 import org.springframework.stereotype.*;
 import org.springframework.transaction.annotation.*;
@@ -7,8 +8,7 @@ import java.util.*;
 import javax.persistence.*;
 
 /**
- * Created by xgono
- * 
+ *
  * @author xgono
  */
 @Repository
@@ -20,18 +20,27 @@ public class TroopDaoImpl implements TroopDao {
     
     @Override
     public Troop create(Troop troop) {
+        if(troop == null) {
+            throw new IllegalArgumentException("troop");
+        }
         em.persist(troop);
         return troop;
     }
 
     @Override
     public Troop update(Troop troop) {
+        if(troop == null) {
+            throw new IllegalArgumentException("troop");
+        }
         em.merge(troop);
         return troop;
     }
 
     @Override
     public Troop delete(Troop troop) {
+        if(troop == null) {
+            throw new IllegalArgumentException("troop");
+        }
         em.remove(em.contains(troop) ? troop : em.merge(troop));
         return troop;
     }
@@ -48,6 +57,9 @@ public class TroopDaoImpl implements TroopDao {
     
     @Override
     public Troop findByName(String name) {
+        if(name == null) {
+            throw new IllegalArgumentException("name");
+        }
         try {
             return em
                     .createQuery("select t from Troop t where name = :name",
